Tanishq_jewelry_recomm_system/requirements.txt
# requirements.txt
torch>=2.0.0
torchvision>=0.15.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
pyarrow>=7.0.0
matplotlib>=3.5.0
Pillow>=9.0.0
tqdm>=4.60.0
ipywidgets>=7.7.0
gdown>=4.5.0
gradio>=3.0.0
concurrent-log-handler>=0.9.20
plotly>=5.10.0
requests
